+++ bitcoin/src/consensus/mod.rs
@@ -8,11 +8,17 @@
 
 pub mod encode;
 pub mod params;
+#[cfg(feature = "bitcoinconsensus")]
+pub mod validation;
 
 pub use self::encode::{
     deserialize, deserialize_partial, serialize, Decodable, Encodable, ReadExt, WriteExt,
 };
 pub use self::params::Params;
+#[cfg(feature = "bitcoinconsensus")]
+pub use self::validation::{
+    verify_script, verify_script_with_flags, verify_transaction, verify_transaction_with_flags,
+};
 
 #[cfg(feature = "serde")]
 pub mod serde;
